Inscripcion andando

+++ src/proyectogrupo67/entidades/Inscripcion.java
@@ -16,19 +16,19 @@ public class Inscripcion {
     private int idInscripcion;
     private Alumno alumno;
     private Materia materia;
-    private Double nota;
+    private int nota;
 
     public Inscripcion() {
     }
 
-    public Inscripcion(int idInscripcion, Alumno alumno, Materia materia, Double nota) {
+    public Inscripcion(int idInscripcion, Alumno alumno, Materia materia, int nota) {
         this.idInscripcion = idInscripcion;
         this.alumno = alumno;
         this.materia = materia;
         this.nota = nota;
     }
 
-    public Inscripcion(Alumno alumno, Materia materia, Double nota) {
+    public Inscripcion(Alumno alumno, Materia materia, int nota) {
         this.alumno = alumno;
         this.materia = materia;
         this.nota = nota;
@@ -58,11 +58,11 @@ public class Inscripcion {
         this.materia = materia;
     }
 
-    public Double getNota() {
+    public int getNota() {
         return nota;
     }
 
-    public void setNota(Double nota) {
+    public void setNota(int nota) {
         this.nota = nota;
     }
